+++ example/auth_example/lib/auth.dart
@@ -7,13 +7,19 @@ class AuthService {
   static late final String _key;
 
   factory AuthService({required String keyname}) {
-    _key = keyname;
+    // Prevent multiple instances with different keynames
+    assert(_key == keyname,
+        'Cannot instantiate more than once with different keyname');
     return _instance;
+  }
+
+  static void initialize({required String keyname}) {
+    _key = keyname;
   }
 
   AuthService._internal() {
     // ignore: unused_local_variable
-    final (usr, setUser) = useStore(_key, UserRuntime());
+    final (usr, _, setUser) = useStore(_key, UserRuntime());
 
     _auth.authStateChanges().listen((User? user) async {
       setUser((value) => UserRuntime(user: user));