+++ lib/o.dart
@@ -1,5 +1,8 @@
 library o;
 
 export 'src/core/app/app_state.dart';
-export 'src/utils/local/local_utils.dart';
+export 'src/utils/local/extensions.dart';
+export 'src/utils/local/use_state.dart';
+export 'src/utils/local/use_t_state.dart';
+export 'src/widgets/t_widget.dart';
 export 'src/widgets/widget.dart';