+++ lib/src/core/app/app_state.dart
@@ -1,8 +1,21 @@
-// GSSM core
-import 'package:flutter/material.dart';
-import 'package:o/src/core/core.dart';
-import 'package:o/src/typedefs/typedefs.dart';
+import 'dart:collection';
 
+import 'package:o/src/core/core.dart';
+
+///
+/// `Store` singleton which holds the AppState.
+///
+/// States are stored within <String, Observable<dynamic>> `_cache`.
+///
+/// exposes 4 **methods**
+/// >-  `addObservable()`
+/// >-  `observable()`
+/// >-  `memoize()`
+/// >-  `removeObservable()`
+///
+/// use `useStore()` to get / set states.
+///
+/// direct access via `getStore()` global function
 class Store {
   static final Store _instance = Store._internal();
 
@@ -12,45 +25,63 @@ class Store {
 
   Store._internal();
 
-  final Map<String, dynamic> _cache = {};
+  final _cache = HashMap<String, Observable<dynamic>>();
 
-  // Observables
+  /// add Observables
   void addObservable<T>({required String key, required T value}) {
     _cache.putIfAbsent(key, () => Observable<T>(value));
   }
 
   Observable<T> observable<T>({required String key, required T initialValue}) {
-    final cachedValue = _cache[key] as Observable<T>?;
-
-    if (cachedValue != null) {
-      return cachedValue;
-    } else {
-      final value = Observable(initialValue);
-      _cache[key] = value;
-      return value;
-    }
+    var cachedValue =
+        _cache.containsKey(key) ? _cache[key] as Observable<T> : null;
+    return cachedValue ??= _cache[key] = Observable<T>(initialValue);
   }
 
-  // Memoization
+  /// Memoization
   Observable<T> memoize<T>(String key, T Function() create) {
-    final cachedValue = _cache[key] as Observable<T>?;
-    if (cachedValue != null) {
-      return cachedValue;
-    } else {
-      final value = Observable(create());
-      _cache[key] = value;
-      return value;
-    }
+    var cachedValue =
+        _cache.containsKey(key) ? _cache[key] as Observable<T> : null;
+
+    return cachedValue ??= _cache[key] = Observable<T>(create());
   }
 
-  // Removal
+  /// Removal
   void removeObservable<T>({required String key}) {
     _cache.remove(key);
   }
 }
 
-// Global store accessor
-(Observable<T> x, void Function(dynamic Function(T)) y) useStore<T>(
+/// Creates an observable store for the provided key and initial value.
+///
+/// Record to expose `Observable` and `void Function(T)` / `void Function(T Function(T))` to create an Appstate with provided
+/// [initialValue] & [key].
+///
+/// ```dart
+/// final (down, setDown, _) = useStore('counter', 500);
+/// setDown(1);
+/// ```
+///
+/// **or**
+///
+/// ```dart
+/// final (down, _, setDown) = useStore('counter', 500);
+/// setDown((prev) => prev - 1);
+/// ```
+/// State will be created only if a state doesn't exist with the provided key.
+///
+/// **Records** :
+/// >- `Observable` | Ex : count
+/// >- `void Function(T)`
+/// >- `void Function(T Function(T)`
+///
+/// **Args**
+/// >> ```dart
+/// String key  // Unique State Identifier
+/// T initialValue
+/// > ```
+(Observable<T> x, void Function(T) y, void Function(T Function(T prev)) z)
+    useStore<T>(
   String key,
   T initialValue,
 ) {
@@ -58,30 +89,10 @@ class Store {
       Store._instance.observable<T>(key: key, initialValue: initialValue);
   return (
     observable,
-    (modifier) => observable.value = modifier(observable.value),
+    (modifier) => observable.value = modifier,
+    (modifier) => observable.value = modifier(observable.value)
   );
 }
 
-// Global Observable injector
-Widget dependencies(List<Tuple> initializers, {required Widget app}) {
-  for (int i = 0; i < initializers.length; i++) {}
-  return app;
-}
-
-// Global service injector
-Widget injector(
-    {required Widget app,
-    List<Object> objects = const [],
-    List<Function> functions = const []}) {
-  for (int i = 0; i < objects.length; i++) {}
-
-  for (int i = 0; i < functions.length; i++) {
-    Function initializer = functions[i];
-    initializer();
-  }
-
-  return app;
-}
-
-// Direct store accessor
+/// Direct store accessor
 Store getStore() => Store._instance;