+++ lib/src/typedefs/typedefs.dart
@@ -1,7 +1,0 @@
-// Type
-import 'package:o/src/core/core.dart';
-
-typedef Tuple<T> = (
-  Observable<T> x,
-  void Function(dynamic Function(T)) y,
-);