+++ test/o_test.dart
@@ -4,23 +4,21 @@ import 'package:o/o.dart';
 
 void main() {
   test('useState test', () {
-    final (count, setCount) = useState(5);
-    setCount((p0) => 5);
+    final (count, setCount, _) = useState(5);
+    setCount(5);
     expect(count.value, 5);
   });
 
   test('useStore test', () {
-    final (gCount, setGCount) = useStore('count', 2);
+    final (gCount, _, setGCount) = useStore('count', 2);
     setGCount((p0) => p0 + 1);
     expect(gCount.value, 3);
 
     getStore().removeObservable(key: 'count');
-    // ignore: unused_local_variable
-    final (newCount, setNewCount) = useStore('count', 100);
-    expect(newCount.value, 100);
+    final (newCount, _, _) = useStore('count', 5);
+    expect(newCount.value, 5);
 
-    // ignore: unused_local_variable
-    final (sameCount, setSameCount) = useStore('count', 500);
-    expect(sameCount.value, 100);
+    final (sameCount, _, _) = useStore('count', 500);
+    expect(sameCount.value, 5);
   });
 }