+++ CHANGELOG.md
@@ -5,4 +5,13 @@
 ## 0.0.2
 
 * Removed unnecessary statements.
-* Renamed globalServices -> injector
+* Renamed globalServices -> injector.
+
+## 0.0.3
+
+* Added optional previous state.
+* Performance optimizations.
+* Added TaggedObserver to filter-out unwanted rebuilds.
+* Removed Dep | Service Injector.
+* Added tag_example and updated main | auth_example.
+* Updated API Docs.